Move TextAlignVertical definition to place accessible by material and cupertino
