+++ packages/flutter/lib/src/cupertino/text_field.dart
@@ -5,7 +5,6 @@ import 'package:flutter/gestures.dart';
 import 'package:flutter/rendering.dart';
 import 'package:flutter/services.dart';
 import 'package:flutter/widgets.dart';
-import 'package:flutter/material.dart' show TextAlignVertical;
 
 import 'colors.dart';
 import 'icons.dart';
