+++ packages/flutter/lib/src/material/input_decorator.dart
@@ -3525,44 +3525,3 @@ class InputDecorationTheme extends Diagnosticable {
     properties.add(DiagnosticsProperty<bool>('alignLabelWithHint', alignLabelWithHint, defaultValue: defaultTheme.alignLabelWithHint));
   }
 }
-
-/// The vertical alignment of text within an input.
-///
-/// A single [y] value that can range from -1.0 to 1.0. -1.0 aligns to the top
-/// of the input so that the top of the first line of text fits within the input
-/// and its padding. 0.0 aligns to the center of the input. 1.0 aligns so that
-/// the bottom of the last line of text aligns with the bottom interior edge of
-/// the input.
-///
-/// See also:
-///
-///  * [TextField.textAlignVertical], which is passed on to the [InputDecorator].
-///  * [CupertinoTextField.textAlignVertical], which behaves in the same way as
-///    the parameter in TextField.
-///  * [InputDecorator.textAlignVertical], which defines the alignment of
-///    prefix, input, and suffix, within the [InputDecorator].
-class TextAlignVertical {
-  /// Construct TextAlignVertical from any given y value.
-  const TextAlignVertical({
-    @required this.y,
-  }) : assert(y != null),
-       assert(y >= -1.0 && y <= 1.0);
-
-  /// A value ranging from -1.0 to 1.0 that defines the topmost and bottommost
-  /// locations of the top and bottom of the input text box.
-  final double y;
-
-  /// Aligns a TextField's input Text with the topmost location within the
-  /// TextField.
-  static const TextAlignVertical top = TextAlignVertical(y: -1.0);
-  /// Aligns a TextField's input Text to the center of the TextField.
-  static const TextAlignVertical center = TextAlignVertical(y: 0.0);
-  /// Aligns a TextField's input Text with the bottommost location within the
-  /// TextField.
-  static const TextAlignVertical bottom = TextAlignVertical(y: 1.0);
-
-  @override
-  String toString() {
-    return '$runtimeType(y: $y)';
-  }
-}
